+++ server/belga/io/feed_parsers/__init__.py
@@ -1,3 +1,7 @@
 from . import belga_newsml_1_2  # noqa
+from . import belga_afp_newsml_1_2  # noqa
+from . import belga_anp_newsml_1_2  # noqa
+from . import belga_efe_newsml_1_2  # noqa
+from . import belga_tass_newsml_1_2  # noqa
 from . import belga_iptc7901  # noqa
 from . import belga_anpa  # noqa
