+++ server/belga/io/feed_parsers/base_belga_newsml_1_2.py
@@ -111,7 +111,7 @@ class BaseBelgaNewsMLOneFeedParser(NewsMLOneFeedParser):
             item['products'] = []
             for element in elements:
                 if element.get('FormalName'):
-                    item['products'].append(element['FormalName'])
+                    item['products'].append(element.get('FormalName', ''))
 
         element = envelop_el.find('Priority')
         if element is not None:
