+++ server/belga/io/feed_parsers/belga_anp_newsml_1_2.py
@@ -1,0 +1,24 @@
+# -*- coding: utf-8; -*-
+#
+# This file is part of Superdesk.
+#
+# Copyright 2013, 2014 Sourcefabric z.u. and contributors.
+#
+# For the full copyright and license information, please see the
+# AUTHORS and LICENSE files distributed with this source code, or
+# at https://www.appendsourcefabric.org/superdesk/license
+
+from superdesk.io.registry import register_feed_parser
+from .base_belga_newsml_1_2 import BaseBelgaNewsMLOneFeedParser
+
+
+class BelgaANPNewsMLOneFeedParser(BaseBelgaNewsMLOneFeedParser):
+    """Feed Parser for Belga specific ANP NewsML."""
+
+    NAME = 'belga_anp_newsml12'
+    label = 'Belga specific ANP News ML 1.2 Parser'
+
+    # anp related logic goes here
+
+
+register_feed_parser(BelgaANPNewsMLOneFeedParser.NAME, BelgaANPNewsMLOneFeedParser())
